Move IsValidHostname out
