+++ go/libkb/util.go
@@ -278,27 +278,6 @@ func IsIn(needle string, haystack []string, ci bool) bool {
 	return false
 }
 
-// Found regex here: http://stackoverflow.com/questions/106179/regular-expression-to-match-dns-hostname-or-ip-address
-var hostnameRE = regexp.MustCompile("^(?i:[a-z0-9]|[a-z0-9][a-z0-9-]*[a-z0-9])$")
-
-func IsValidHostname(s string) bool {
-	parts := strings.Split(s, ".")
-	// Found regex here: http://stackoverflow.com/questions/106179/regular-expression-to-match-dns-hostname-or-ip-address
-	if len(parts) < 2 {
-		return false
-	}
-	for _, p := range parts {
-		if !hostnameRE.MatchString(p) {
-			return false
-		}
-	}
-	// TLDs must be >=2 chars
-	if len(parts[len(parts)-1]) < 2 {
-		return false
-	}
-	return true
-}
-
 func RandBytes(length int) ([]byte, error) {
 	var n int
 	var err error
